+++ spec/www/spec/pre-populated-db-test.js
@@ -4,21 +4,14 @@ var MYTIMEOUT = 12000;
 
 var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
 
-// FUTURE TODO replace in test(s):
-function ok(test, desc) { expect(test).toBe(true); }
-function equal(a, b, desc) { expect(a).toEqual(b); } // '=='
-function strictEqual(a, b, desc) { expect(a).toBe(b); } // '==='
-
-var isAndroid = /Android/.test(navigator.userAgent);
-var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
-var isWindows = /Windows /.test(navigator.userAgent); // Windows (8.1)
-var isIE = isWindows || isWP8;
+var isWindows = /Windows /.test(navigator.userAgent);
+var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
 
 var pluginScenarioList = [ isAndroid ? 'Plugin-sqlite-connector' : 'Plugin', 'Plugin-android.database' ];
 
-var pluginScenarioCount = isAndroid ? 2 : 1;
+//var pluginScenarioCount = isAndroid ? 2 : 1;
+var pluginScenarioCount = 1;
 
-// simple tests:
 var mytests = function() {
 
   for (var i=0; i<pluginScenarioCount; ++i) {
@@ -31,13 +24,14 @@ var mytests = function() {
       it(suiteName + 'preliminary cleanup',
         function(done) {
           expect(true).toBe(true);
-          window.sqlitePlugin.deleteDatabase('pre.db', done, done);
+          window.sqlitePlugin.deleteDatabase({name: 'pre.db', location: 0}, done, done);
         }, MYTIMEOUT);
 
       it(suiteName + 'Pre-populated database test',
         function(done) {
           var dbc1 = window.sqlitePlugin.openDatabase({
             name: 'pre.db',
+            location: 0,
             createFromLocation: 1,
             androidDatabaseImplementation: isOldDatabaseImpl ? 2 : 0
           });
@@ -55,8 +49,7 @@ var mytests = function() {
               check1 = true;
 
               // try some changes:
-              tx.executeSql('DROP TABLE tt');
-              tx.executeSql('CREATE TABLE tt (testcol)');
+              tx.executeSql('DELETE FROM tt');
               tx.executeSql('INSERT INTO tt VALUES (?)', ['new-value']);
             });
           }, function(e) {
@@ -69,6 +62,7 @@ var mytests = function() {
               // try opening it again:
               var dbc2 = window.sqlitePlugin.openDatabase({
                 name: 'pre.db',
+                location: 0,
                 createFromLocation: 1,
                 androidDatabaseImplementation: isOldDatabaseImpl ? 2 : 0
               });
@@ -95,6 +89,7 @@ var mytests = function() {
             });
           });
         }, MYTIMEOUT);
+
     });
   };
 }